+++ src/components/molecules/FilesInput/index.tsx
@@ -3,7 +3,7 @@ import { useField } from 'formik'
 import { toast } from 'react-toastify'
 import FileInfo from './Info'
 import FileInput from './Input'
-import { getFileInfo } from '../../../utils'
+import getFileInfo from '../../../utils/file'
 import { InputProps } from '../../atoms/Input'
 
 interface Values {
